requirements.txt
Flask-Bootstrap==3.3.7.1
Pillow==7.2.0
line-bot-sdk==1.13.0
pandas==0.25.1
Flask==1.1.2
Flask-Cors==3.0.8
line-bot-sdk==1.13.0
pandas==0.25.1
google-api-core==1.21.0
google-auth==1.18.0
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-cloud-firestore==1.8.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
